backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
pymongo==3.12.3
djongo==1.3.6
dnspython==2.1.0
python-dotenv==0.19.2
django-cors-headers==3.7.0
whitenoise==6.4.0
waitress==2.1.2
